Test button adds value to array given index
The test button can now insert a value at a specific index
in the gameBoard array. The button triggers a function that
compares the button's id to the indices in the gameBoard.
Each clickable game tile will have an id that represents it's
position in the gameBoard. When the tile's id matches the current index
it will insert the player's value.