Ability to toggle players
Defined player_x and player_o with some dummy data.
Now have the ability to togglePlayer after a valid move is made.
might want to call togglePlayer differently in the future.
togglePlayer won't toggle back with current button because button
only targets specific index at the moment and the function won't run
unless it is targeting an index with a value of empty string ''