HTML board clickable and added tie game checker
Added the ability to click a tile and add a moveValue to the
JS gameBoard. Adding value to gameBoard relies on
comparing the id of the clicked tile div to the gameBoard's indices.
Not currently compatible with a gameBoard that could change size
because the ids are hard coded onto the HTML tile divs.
Also added a checkForOver function that will see if the game
is either a tie or a player won. The playerWin condition
is currently set to false for testing the the playersTie function.
playersTie currently console logs a message when playerWin is false
and when no tile in the array is an empty string ''