Created better user Auth feedback
Optimizes HTML to receive messaging from authUi module.
authUi now sends useful feedback to the user for
both success and failure options.
Next steps include having signUp also trigger signIn so that a user
is logged in on sign up and begins a valid user session.
Also need to show and hide certain functionality from the user
depending whether or not they are in a session (signed in).
A user shouldn't see sign out or changePassword
when they have not signed in.